add checkbox for conset
