+++ src/locale/de.json
@@ -97,6 +97,10 @@
         "print": "Diese Seite drucken",
         "report": "Einen Fehler melden"
       },
+      "announcement": {
+        "title": "Ankündigung",
+        "button": "Verstanden"
+      },
       "button": {
         "cancel": "Abbrechen",
         "clear": "Löschen",
